+++ src/types.ts
@@ -3,6 +3,7 @@ import type { DefaultLayoutConfig, DefaultVariants } from './defaultConfig/types
 import type { NotificationPosition } from './types/config'
 import type { GestureConfig } from './types/gestures'
 import type { ComponentProps, VFC } from 'react'
+import type { ModifiedEmitParam } from './core/services/types'
 
 declare global {
   namespace Notificated {
@@ -40,3 +41,18 @@ export type NotificationsConfig<Variants> = {
   DefaultLayoutConfig
 
 export type { CustomAnimationConfig }
+
+export type Modify = (id: string, params: Partial<ModifiedEmitParam>) => void
+
+export type Remove = (id: string) => void
+
+export type Notify<V extends VariantsMap = Variants> = <Variant extends keyof V>(
+  notificationType: Variant,
+  setup: { params: RequiredProps<V[Variant]>; config?: Partial<NotificationConfigBase> }
+) => { id: string }
+
+export type UseNotification<V extends VariantsMap = Variants> = () => {
+  modify: Modify
+  remove: Remove
+  notify: Notify<V>
+}
